Added some logs.

+++ source/pdb_parser_module.py
@@ -4,6 +4,7 @@ Module with object PDB containing all the information.
 from __future__ import absolute_import
 import os
 import urllib
+import logging
 from Bio.PDB import PDBParser
 
 class PDB():
@@ -23,14 +24,30 @@ class PDB():
         """
         Initialization of PDB Object.
         """
+        logging.info("Initializing PDB Object...")
+
+        logging.info("Checking if file name was specified...")
         if arg_file_name == ".":
+            logging.info("File name not specified, it will be the organism name.")
             arg_file_name = arg_organism
+        
+        logging.info("Setting save location to %s", arg_save_location)
         self.save_location = arg_save_location
+
+        logging.info("Setting organism to %s", arg_organism)
         self.organism = arg_organism
+
+        logging.info("Setting file name to %s.pdb", arg_file_name)
         self.file_name = arg_file_name + ".pdb"
+
+        logging.info("Calling make_url function...")
         self.make_url(arg_url)
+
+        logging.info("Calling download_url function...")
         self.download_url()
+
         if not self.failed:
+            logging.info("Process not failed, initializing PDB parser function...")
             self.parse_pdb()
 
     def make_url(self, arg_url):
@@ -39,7 +56,10 @@ class PDB():
         and converts it in to a download link for the PDB file.
         Returns the full URL for download.
         """
+        logging.info("Creating URL...")
         self.url = arg_url + self.organism + ".pdb"
+        logging.info("The URL for download is %s", self.url)
+        
 
     def download_url(self):
         """
@@ -49,29 +69,39 @@ class PDB():
         Return True if the download was successful and False if it wasn't and the real PDB
         file location with his name.
         """
+
+        logging.info("Starting the download procedure...")
+
         file_exists = False
 
-        if os.path.exists(self.save_location):
-            pass
-        else:
+        logging.info("Checking if save location '%s' exists...", self.save_location)
+        if not os.path.exists(self.save_location):
+            logging.info("Save location don't exists, trying to create it...")
             try:
                 os.makedirs(self.save_location)
             except OSError:
+                logging.error("Failed to create the save location '%s' using makedirs. \
+                               Download can't proceed.", self.save_location)
                 self.failed = True
                 file_exists = False
             else:
+                logging.info("Checking if file exists...")
                 if os.path.exists(self.save_location + self.file_name):
+                    logging.warning("File exists, download will not proceed.")
                     file_exists = True
 
-        if not file_exists:
-            self.failed = True
+        if not file_exists and not self.failed:
+            logging.info("Trying to download the file from %s on %s", self.url, self.save_location)
             while self.failed_count < 5 and self.failed:
                 try:
                     urllib.request.urlretrieve(self.url, self.save_location + self.file_name)
                 except urllib.error.URLError:
+                    logging.error("Download failed! Trying again. %s tryies left.", \
+                                 (5 - self.failed_count))
                     self.failed = True
                     self.failed_count += 1
                 else:
+                    logging.info("Download successful in %i tryies.", self.failed_count + 1)
                     self.failed = False
                     self.failed_count = 0
 
@@ -79,9 +109,16 @@ class PDB():
         """
         Takes all the atom information from the PDB and adds it in to a dictionary.
         """
+        logging.info("parse_pdb starting...")
         parser = PDBParser()
+
+        logging.info("Transforming PDB in to an object...")
         structure = parser.get_structure(self.organism, self.save_location + self.file_name)
+
+        logging.info("Parsing ATOM entries...")
         for atom in structure.get_atoms():
+            logging.debug("Parsing atom number %i named %s.", atom.get_serial_number(), \
+                                                              atom.get_name())
             self.pdb_dictionary["ATOM"][atom.get_serial_number()] = {\
             "serial_number" : atom.get_serial_number(),\
             "name" : atom.get_name(),\
